requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.1.3
psycopg2-binary
python-dotenv
